fix: resolve all TypeScript and ESLint errors
